ABC de productos y comboxes

+++ Proyecto Ordinario/Clientes.cs
@@ -15,9 +15,11 @@ namespace Proyecto_Ordinario
     {
         //Conexion a la base de datos
         SqlConnection conexion = new SqlConnection();
-        //string cadena = "Data Source=PC-DAVID;Initial Catalog=TiendaTec;Integrated Security=True";
+        string cadena = "Data Source=PC-DAVID;Initial Catalog=TiendaTec;Integrated Security=True";
         //string cadena = "Data Source=shamiko;Initial Catalog=TiendaTec;Integrated Security=True";
-        string cadena = "Data Source=DESKTOP-445GP77;Initial Catalog=TiendaTec;Integrated Security=True"; 
+        //string cadena = "Data Source=DESKTOP-445GP77;Initial Catalog=TiendaTec;Integrated Security=True"; 
+        int id = 0;
+
         public Clientes()
         {
             InitializeComponent();
@@ -25,7 +27,6 @@ namespace Proyecto_Ordinario
             //conexion.Open();
         }
 
-        int id = 0;
         private void btnSiguiente_Click(object sender, EventArgs e)
         {
             conexion.Open();
@@ -56,40 +57,52 @@ namespace Proyecto_Ordinario
 
         private void btnQuery_Click(object sender, EventArgs e)
         {
+            conexion.Open();
             SqlCommand cmd = new SqlCommand("SELECT * FROM vw_Clientes", conexion);
             SqlDataAdapter adaptador = new SqlDataAdapter();
             adaptador.SelectCommand = cmd;
             DataTable tabla = new DataTable();
             adaptador.Fill(tabla);
             dataGridView1.DataSource = tabla;
+            conexion.Close();
         }
 
         private void btnAgregar_Click(object sender, EventArgs e)
         {
             conexion.Open();
-            int id = Convert.ToInt16(txtId.Text);
-            string cadena = "insert into Clientes values (" + id + ",'"+ txtN.Text + "','" + txtAP.Text + "','" + txtAM.Text + "','" + txtC.Text + "')";
-            SqlCommand comando = new SqlCommand(cadena, conexion);
-            comando.ExecuteNonQuery();
-            MessageBox.Show("Los datos se guardaron correctamente");
+            if (txtId.Text != "" && txtN.Text != "" && txtAP.Text != "" && txtAM.Text != "" && txtC.Text != "")
+            {
+                int id = Convert.ToInt16(txtId.Text);
+                string cadena = "insert into Clientes values (" + id + ",'" + txtN.Text + "','" + txtAP.Text + "','" + txtAM.Text + "','" + txtC.Text + "')";
+                SqlCommand comando = new SqlCommand(cadena, conexion);
+                comando.ExecuteNonQuery();
+                MessageBox.Show("Los datos se guardaron correctamente");
+                Limpiar();
+            }
+            else
+                MessageBox.Show("Lllenar los campos");
             conexion.Close();
-            Limpiar();
         }
 
         private void btnBorrar_Click(object sender, EventArgs e)
         {
             conexion.Open();
-            SqlCommand comando = new SqlCommand("delete from Clientes where Id_Cliente=" + txtId.Text, conexion);
-            int cant;
-            cant = comando.ExecuteNonQuery();
-            if (cant == 1)
+            if (txtId.Text != "")
             {
-                MessageBox.Show("Se borró el artículo");
+                SqlCommand comando = new SqlCommand("delete from Clientes where Id_Cliente=" + txtId.Text, conexion);
+                int cant;
+                cant = comando.ExecuteNonQuery();
+                if (cant == 1)
+                {
+                    MessageBox.Show("Se borró el cliente");
+                }
+                else
+                {
+                    MessageBox.Show("No existe un cliente con el ID ingresado");
+                }
             }
             else
-            {
-                MessageBox.Show("No existe un artículo con el código ingresado");
-            }
+                MessageBox.Show("No dejar vacio el ID");
             conexion.Close();
             Limpiar();
         }
@@ -97,52 +110,68 @@ namespace Proyecto_Ordinario
         private void btnGuardar_Click(object sender, EventArgs e)
         {
             conexion.Open();
-            int id = Convert.ToInt16(txtId.Text);
-            string cadena = "UPDATE Clientes set Id_Cliente='" + id + "',Nombre='" + txtN.Text + "',Ap_Pat='" + txtAP.Text + "',Ap_Mat='" + txtAM.Text + "',Correo='" + txtC.Text + "' WHERE Id_Cliente=" + txtId.Text;
-            SqlCommand comando = new SqlCommand(cadena, conexion);
-            int cant;
-            cant = comando.ExecuteNonQuery();
-            if (cant == 1)
+            if (txtId.Text != "" && txtN.Text != "" && txtAP.Text != "" && txtAM.Text != "" && txtC.Text != "")
             {
-                MessageBox.Show("Se modificaron los datos del artículo");
-                Limpiar();
+                int id = Convert.ToInt16(txtId.Text);
+                string cadena = "UPDATE Clientes set Id_Cliente='" + id + "',Nombre='" + txtN.Text + "',Ap_Pat='" + txtAP.Text + "',Ap_Mat='" + txtAM.Text + "',Correo='" + txtC.Text + "' WHERE Id_Cliente=" + txtId.Text;
+                SqlCommand comando = new SqlCommand(cadena, conexion);
+                int cant;
+                cant = comando.ExecuteNonQuery();
+                if (cant == 1)
+                {
+                    MessageBox.Show("Se modificaron los datos del cliente");
+                    Limpiar();
+                }
+                else
+                    MessageBox.Show("No existe un cliente con el ID ingresado");
             }
             else
-                MessageBox.Show("No existe un artículo con el código ingresado");
+                MessageBox.Show("Lllenar los campos");
             conexion.Close();
         }
 
         private void btnModificar_Click(object sender, EventArgs e)
         {
             conexion.Open();
-            string cod = txtId.Text;
-            string cadena = "select * from Clientes where Id_Cliente =" + cod;
-            SqlCommand comando = new SqlCommand(cadena, conexion);
-            SqlDataReader registro = comando.ExecuteReader();
-            if (registro.Read())
+            if (txtId.Text != "")
             {
-                txtId.Text = registro["Id_Cliente"].ToString();
-                txtN.Text = registro["Nombre"].ToString();
-                txtAP.Text = registro["Ap_Pat"].ToString();
-                txtAM.Text = registro["Ap_Mat"].ToString();
-                txtC.Text = registro["Correo"].ToString();
+                string cod = txtId.Text;
+                string cadena = "select * from Clientes where Id_Cliente =" + cod;
+                SqlCommand comando = new SqlCommand(cadena, conexion);
+                SqlDataReader registro = comando.ExecuteReader();
+                if (registro.Read())
+                {
+                    txtId.Text = registro["Id_Cliente"].ToString();
+                    txtN.Text = registro["Nombre"].ToString();
+                    txtAP.Text = registro["Ap_Pat"].ToString();
+                    txtAM.Text = registro["Ap_Mat"].ToString();
+                    txtC.Text = registro["Correo"].ToString();
+                }
+                else
+                    MessageBox.Show("No existe un cliente con el id ingresado");
             }
             else
-                MessageBox.Show("No existe un cliente con el id ingresado");
+                MessageBox.Show("No dejar vacio el ID");
             conexion.Close();
         }
 
         private void BtnBuscar_Click(object sender, EventArgs e)
         {
             conexion.Open();
-            SqlCommand cmd = new SqlCommand("sp_NombresClientes ", conexion);
-            cmd.CommandType = CommandType.StoredProcedure;
-            cmd.Parameters.AddWithValue("@Id", txtId.Text);
-            DataTable dt = new DataTable();
-            dt.Load(cmd.ExecuteReader());
-            dataGridView1.DataSource = dt;
+            if (txtId.Text != "")
+            {
+                SqlCommand cmd = new SqlCommand("sp_NombresClientes ", conexion);
+                cmd.CommandType = CommandType.StoredProcedure;
+                cmd.Parameters.AddWithValue("@Id", txtId.Text);
+                DataTable dt = new DataTable();
+                dt.Load(cmd.ExecuteReader());
+                dataGridView1.DataSource = dt;
+                conexion.Close();
+                Limpiar();
+            }
+            else
+                MessageBox.Show("No dejar vacio el ID");
             conexion.Close();
-            Limpiar();
         }
 
         private void btnPrimero_Click(object sender, EventArgs e)
@@ -159,7 +188,14 @@ namespace Proyecto_Ordinario
 
         private void btnUltimo_Click(object sender, EventArgs e)
         {
-
+            conexion.Open();
+            SqlCommand cmd = new SqlCommand("SELECT TOP 1 * FROM Clientes ORDER BY Id_Cliente DESC", conexion);
+            SqlDataAdapter adaptador = new SqlDataAdapter();
+            adaptador.SelectCommand = cmd;
+            DataTable tabla = new DataTable();
+            adaptador.Fill(tabla);
+            dataGridView1.DataSource = tabla;
+            conexion.Close();
         }
 
         private void Limpiar()
